requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.30.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
